Init
Doing this commit even before jam has ended. Yeah, game didn't turned out great, but I am glad that I learned a lot of cool stuff on the way. It would be greate to add sounds and scoring features, but project surely needs refactoring to make it enjoyable to work on it again. It was a great first try to make a complete gamein c++ and SDL2)